+++ pyproject.toml
@@ -30,7 +30,6 @@ aiohttp = ">=3.0.0"
 yarl = ">=1.6.0"
 
 [tool.poetry.group.dev.dependencies]
-aresponses = "2.1.6"
 codespell = "2.2.6"
 covdefaults = "2.3.0"
 coverage = {version = "7.3.3", extras = ["toml"]}
@@ -45,6 +44,7 @@ ruff = "0.1.8"
 safety = "2.4.0b2"
 yamllint = "1.33.0"
 syrupy = "4.6.0"
+aioresponses = "^0.7.6"
 
 [tool.poetry.urls]
 "Bug Tracker" = "https://github.com/joostlek/python-withings/issues"
